+++ CHANGELOG.md
@@ -6,7 +6,11 @@
 
 ### Added
 
+- Add support for platform version `242.*`
+
 ### Changed
+
+- `pluginVerifierIdeVersions` – upgrade to `2022.3.3, 2023.1.6, 2023.2.6, 2023.3.6, 2024.1.1, 2024.2`
 
 ### Deprecated