+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 
 ### Fixed
 
+- Refresh `ErrorLogTreeModel` on AWT event dispatching thread
+
 ### Security
 
 ## [1.4.5] - 2024-01-20