+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 
 ### Security
 
-## 1.4.6 - 2024-05-19
+## [1.4.6] - 2024-05-19
 
 ### Added
 
@@ -601,6 +601,7 @@
   - Up vector axis selection
 
 [Unreleased]: https://github.com/sczerwinski/wavefront-obj-intellij-plugin/compare/v1.4.6...main
+[1.4.6]: https://github.com/sczerwinski/wavefront-obj-intellij-plugin/compare/v1.4.5...v1.4.6
 [1.4.5]: https://github.com/sczerwinski/wavefront-obj-intellij-plugin/compare/v1.4.4...v1.4.5
 [1.4.4]: https://github.com/sczerwinski/wavefront-obj-intellij-plugin/compare/v1.4.3...v1.4.4
 [1.4.3]: https://github.com/sczerwinski/wavefront-obj-intellij-plugin/compare/v1.4.2...v1.4.3